Rename and modularize makeCanvasAutoFullscreen to fitOnResize
